+++ CHANGELOG.md
@@ -5,12 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-
 ## [Unreleased]
 
 ### Changed
 
 - Include k8s dependency for 1.18 in dependabot
+
+## Security
+
+- Update actions/cache actions/upload-artifact to v2.1.1 in generated
+  workflows.
 
 ## [1.0.0] - 2020-09-23